fix(irrigation): portalApLoop usa unsigned long (evita mismatch de mangling uint32_t)
Pré-existente (não relacionado à supressão meteorológica): uint32_t resolvia como
unsigned int em PortalAp.cpp (via DNSServer.h/WiFi.h) e unsigned long no TU do chamador
(IrrigationModule.cpp), gerando 'undefined reference to portalApLoop(unsigned long)' no
link ESP32. Assinatura explícita unsigned long (domínio de millis()) elimina a ambiguidade.

+++ src/modules/irrigation/PortalAp.cpp
@@ -42,7 +42,7 @@ static void tearDown()
     sApUp = false;
 }
 
-void portalApLoop(uint32_t nowMs)
+void portalApLoop(unsigned long nowMs)
 {
     if (!irrigationModule)
         return;
